# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "maklaGo API",
    "version": "1.0.0",
    "description": "Restaurant POS backend. **Tenant**: optional header `x-tenant-id` (slug or registry UUID). With `Authorization: Bearer` staff JWT, the tenant is resolved from the token when the header is omitted. **Language**: optional header **`lang`** (`fr`, `en`, `ar`; default **`fr`**) on every request — error messages and localized strings follow it; response includes **`Content-Language`**. **`POST /api/v1/auth/login`** accepts optional `x-tenant-id` for tenant-scoped login; without it, the normalized email must exist in registry **`staff_login_directory`** (slugless directory login). **`GET /api/v1/orders`** returns enriched catalog labels on lines (dashboard); **`GET /api/v1/mobile/orders`** uses a slimmer shell than **`POST /orders`**: no **`staff`**, no nested **`table`** object; dine-in adds root **`tableId`** + **`tableNumber`**. Lines add **`products[].name`**, **`extras[].name`**, optional **`extras[].typeId`**, optional **`compositionSnapshot`**. **`POST /api/v1/orders`** accepts **`products`** snapshots (`categoryId`, `id`, `count`, **`price`**, **`extras`**) for analytics; **`subtotal`**, **`tva`**, **`total`** use major currency units in JSON (e.g. 2.5); DB keeps cents internally (must match server); new orders start as **`waiting`**. **Platform** routes use `x-platform-key`. Errors: `{ \"error\": { \"code\", \"message\", \"details?\" } }`."
  },
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Staff JWT from `/api/v1/auth/login`."
      },
      "PlatformKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-platform-key",
        "description": "Must match `PLATFORM_API_KEY` on the server."
      }
    },
    "schemas": {
      "Currency": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "curr_tnd" },
          "code": { "type": "string", "example": "TND" },
          "name": { "type": "string", "example": "Dinar Tunisien" },
          "symbol": { "type": "string", "example": "DT" }
        }
      },
      "RestaurantSettings": {
        "type": "object",
        "properties": {
          "tva": { "type": "number", "example": 19, "description": "Global TVA rate as percentage (e.g. 19 = 19%). Use this when computing order totals." },
          "openTime": { "type": "string", "example": "12:00", "description": "Business day start (HH:mm, 24h)." },
          "closeTime": { "type": "string", "example": "01:00", "description": "Business day end (HH:mm, 24h, may be past midnight)." },
          "currency": { "$ref": "#/components/schemas/Currency" }
        }
      },
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
      "PatchTaxRequest": {
        "type": "object",
        "required": [
          "tax"
        ],
        "properties": {
          "tax": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100000
          }
        }
      },
...
    },
    "parameters": {
...
      "TenantHeaderOptional": {
        "name": "x-tenant-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Tenant slug (e.g. demo) or registry tenant UUID. Optional when the request includes a valid staff Bearer JWT—the server resolves tenant from the token."
      },
...
    }
  },
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "All tenant settings (KV map)",
        "description": "Returns the full settings key-value map plus a parsed `tax`. Known keys: `default_tax_bps` (integer), `open_time` (string HH:mm), `close_time` (string HH:mm), `currency_id` (string). Use the typed endpoints (`/settings/restaurant`, `/settings/tax`) to read or write individual settings.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "description": "Raw KV map — keys vary by what has been written.",
                      "example": {
                        "default_tax_bps": 1000,
                        "open_time": "12:00",
                        "close_time": "01:00",
                        "currency_id": "curr_tnd"
                      }
                    },
                    "tax": {
                      "type": "integer",
                      "description": "Parsed default TVA rate in basis points (100 bps = 1%).",
                      "example": 1000
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/tax": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set default tax (basis points)",
        "description": "Requires role owner or manager.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTaxRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          }
        ]
      }
    },
    "/api/v1/settings/restaurant": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get restaurant settings",
        "description": "Returns TVA rate (%), operating hours, and display currency. Defaults: 19% TVA, 12:00–01:00, TND. **Mobile apps should call `GET /api/v1/mobile/config` instead.**",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RestaurantSettings" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Settings"],
        "summary": "Update restaurant settings (owner/manager)",
        "description": "All fields optional. `openTime` / `closeTime` must be `HH:mm` format. `tva` is a percentage (e.g. `19` for 19%). Use `GET /api/v1/settings/currencies` to list valid `currencyId` values.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "openTime": { "type": "string", "pattern": "^\\d{2}:\\d{2}$", "example": "12:00" },
                  "closeTime": { "type": "string", "pattern": "^\\d{2}:\\d{2}$", "example": "01:00" },
                  "currencyId": { "type": "string" },
                  "tva": { "type": "number", "minimum": 0, "maximum": 100, "example": 19, "description": "TVA rate as percentage (e.g. 19 = 19%). Stored internally as basis points." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestaurantSettings" } } }
          },
          "400": {
            "description": "validation_error — invalid format",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "403": {
            "description": "Owner or manager only",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          }
        }
      }
    },
...
    "/api/v1/settings/currencies": {
      "get": {
        "tags": ["Settings"],
        "summary": "List available currencies",
        "description": "Seeded with USD ($), EUR (€), TND (DT). Use the `id` as `currencyId` in PATCH /api/v1/settings/restaurant.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "currencies": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Currency" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Create a currency (owner/manager)",
        "description": "Creates a new currency. `code` must be unique (stored uppercase).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/TenantHeaderOptional" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code", "name", "symbol"],
                "properties": {
                  "code": { "type": "string", "maxLength": 10, "example": "GBP" },
                  "name": { "type": "string", "maxLength": 100, "example": "Pound Sterling" },
                  "symbol": { "type": "string", "maxLength": 10, "example": "£" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Currency created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "currency": { "$ref": "#/components/schemas/Currency" } }
                }
              }
            }
          },
          "400": {
            "description": "validation_error — missing or invalid fields",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "403": {
            "description": "Owner or manager only",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "409": {
            "description": "currency_code_taken — code already exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          }
        }
      }
    },
    "/api/v1/settings/currencies/{id}": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get a currency by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "currency": { "$ref": "#/components/schemas/Currency" } }
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          }
        }
      },
      "patch": {
        "tags": ["Settings"],
        "summary": "Update a currency (owner/manager)",
        "description": "All fields optional. `code` must remain unique.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": { "type": "string", "maxLength": 10 },
                  "name": { "type": "string", "maxLength": 100 },
                  "symbol": { "type": "string", "maxLength": 10 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated currency",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "currency": { "$ref": "#/components/schemas/Currency" } }
                }
              }
            }
          },
          "400": {
            "description": "validation_error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "403": {
            "description": "Owner or manager only",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "404": {
            "description": "not_found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "409": {
            "description": "currency_code_taken — code already exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          }
        }
      },
      "delete": {
        "tags": ["Settings"],
        "summary": "Delete a currency (owner/manager)",
        "description": "Fails with 409 if this currency is the active restaurant currency. Change the restaurant currency first, then retry.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "403": {
            "description": "Owner or manager only",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "404": {
            "description": "not_found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          },
          "409": {
            "description": "currency_in_use — cannot delete the active currency",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
          }
        }
      }
    }
  }
}